test: move to ecrecover.rs
